rename interface folder to HardwareInterfaces
rename HardwareInterface file to hardware
abstract.

update ini to reflect new file structure
