+++ Platformio/HAL/HardwareAbstract.hpp
