+++ Platformio/HAL/HardwareInterfaces/BatteryInterface.h
