+++ Platformio/platformio.ini
@@ -40,14 +40,14 @@ build_flags =
   ; ------------- Includes ------------------------------------
   -I OmoteUI
   -I HAL
-  -I HAL/Interface
+  -I HAL/HardwareInterfaces
 
 lib_deps =
   lvgl/lvgl@^8.3.9
 lib_archive = false
 build_src_filter =
   +<../OmoteUI/*>
-  +<../HAL/Interface/*>
+  +<../HAL/HardwareAbstract.cpp>
